+++ data/db/build.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS guilds(
     LevelMessages text DEFAULT "no"
 );
 
-CREATE TABLE IF NOT EXISTS exp (
+CREATE TABLE IF NOT EXISTS exp(
     UserID integer PRIMARY KEY,
     XP integer DEFAULT 0,
     Level integer DEFAULT 0,
@@ -16,8 +16,8 @@ CREATE TABLE IF NOT EXISTS exp (
 
 CREATE TABLE IF NOT EXISTS mutes(
     UserID integer PRIMARY KEY,
-    RoleIDs text,
-	GuildID integer
+	GuildID integer SECONDARY KEY,
+    RoleIDs text
 );
 
 CREATE TABLE IF NOT EXISTS votes(